+++ src/uses.h
@@ -73,7 +73,10 @@ SOFTWARE.
 #if    defined(USE_CMD_LN_S) \
     || defined(USE_CMD_LN_H) \
     || defined(USE_CMD_MV) \
-    || defined(USE_CMD_EXPORT)
+    || defined(USE_CMD_EXPORT) \
+    || defined(USE_CMD_WHILE_NO_ERROR) \
+    || defined(USE_CMD_WHILE_ERROR) \
+    || defined(USE_CMD_FOR_EACH)
 
 #define USES_SHARED_STR
 
@@ -107,23 +110,27 @@ SOFTWARE.
 
 // USES_ENVIRONMENT - any command or capability that performs management of environment variables.
 #if    defined(USE_ENVIROMENT_INPUT) \
-    || defined(USE_CMD_SPAWN) \
-    || defined(USE_CMD_PWD) \
-    || defined(USE_CMD_WAIT_PID) \
     || defined(USE_CMD_ENV_CAT_FD)
 
 #define USES_ENVIRONMENT
+
+// Even though these use setenv, it itself doesn't require
+//   the environment parser.
+//    || defined(USE_CMD_EXPORT)
+//    || defined(USE_CMD_SPAWN)
+//    || defined(USE_CMD_PWD)
+//    || defined(USE_CMD_WAIT_PID)
 
 #endif
 
 
 // USES_SHARED_SPLIT_ARG - shared data for splitting one arg into multiple args.
 #if    defined(USE_CMD_EXEC) \
-    || defined(USE_CMD_EXPORT) \
     || defined(USE_CMD_SPAWN) \
     || defined(USE_CMD_WAIT_PID) \
     || defined(USE_CMD_SUBCMD) \
-    || defined(USE_CMD_IF_ELSE)
+    || defined(USE_CMD_IF_ELSE) \
+    || defined(USE_CMD_FOR_EACH)
 
 #define USES_SHARED_SPLIT_ARG